chore(clerk-js): Remove FileDropArea
